fix: [DHIS2-12112] add missing programStage rules
